Create servlet to display single view
